Packager dss.py updated
Projection WKT set to HEC's WKT from AlbersInfo Java class.
The pydsstools package uses a grid transform defined with Affine
and that grid transform definition is set to zeros except postion
zero (0) as the cellsize.  This, along with a mode to grid.pyx,
sets the XY Coord of grid cell to zero; what CWMS CAVI 3.2.1 requires.