NDFD CONUS products for QPF and Air Temperature
NDFD 1-3 day and 4-7 day QPF and Air Temperature added.
Both 1-3 and 4-7 day QPF as a single acquirable due to
the time interval being the same at 6-hours.  Air Temperature
for the 1-3 day is mostly 1-hour with the last six grids at
a 3-hour interval.  Day 4-7 for Air Temperature is 6-hour.