Packager Python config and Dockerfile
- Python setup config mod removing GDAL<=3.5 requirement
- Slight refactor Dockerfile to use `tiffdss` C library release

The Python package setup.cfg file defined the GDAL version was to
be less than or equal to version 3.5.0; therefore, restricting version
3.6.  That requirement has been removed to allow the base Docker image
to define the GDAL version.

The Packager Dockerfile originally cloned HEC-DSS, Tiffdss, and compiled
the C library for tiffdss (libtiffdss.so) that is used in the DSS7
packager to write records from GeoTiff to DSS7.  The Tiffdss GitHub repo
now provides the compiled shared object in a release eliminating the
need for the packager to compile on build.